# Improve performance of quantizelinear for int4